+++ changelog.md
@@ -8,6 +8,7 @@ improvements may be omitted.)_
 
 __[maintenance]__
 
+* Update certificate information for Blender LTS installer.
 * NLog library is updated from 5.3.4 to 5.4.0.
 
 ## Version 2025.01.31.0